--- a/GNUmakefile
+++ b/GNUmakefile
@@ -1,4 +1,5 @@
 GOFMT_FILES?=$$(find . -name '*.go' |grep -v vendor)
+KAFKA_BOOTSTRAP_SERVERS ?= localhost:9092
 default: build
 
 build:
@@ -9,7 +10,7 @@ test:
 
 testacc:
 	GODEBUG=x509ignoreCN=0 \
-	KAFKA_BOOTSTRAP_SERVERS=localhost:9092 \
+	KAFKA_BOOTSTRAP_SERVERS=$(KAFKA_BOOTSTRAP_SERVERS) \
 	KAFKA_CA_CERT=../secrets/ca.crt \
 	KAFKA_CLIENT_CERT=../secrets/client.pem \
 	KAFKA_CLIENT_KEY=../secrets/client.key \