When creating/destroying 1000 ACLs:

### On a real cluster with ~80k ACLs

#### Provides 89000% speedup for ACL creations (~8000s -> ~9s)

### On a fresh cluster with no ACLs yet: 

#### Provides a 780% speedup for ACL creations and 1470% speedup for ACL deletions .

Before:
```shell
time terraform apply -auto-approve -parallelism=1000 
25.98s user 18.43s system 83% cpu 53.286 total

time terraform destroy -parallelism=1000 -auto-approve  
39.98s user 19.88s system 32% cpu 3:04.59 total

```

After:
```shell
time terraform apply -parallelism=1000 -auto-approve
25.81s user 15.20s system 603% cpu 6.798 total

time terraform destroy -parallelism=1000 -auto-approve 
30.12s user 13.22s system 344% cpu 12.570 total
```


For anyone facing the same issue, I've published a release here until this PR is reviewed/merged:
https://registry.terraform.io/providers/eh-steve/kafka/latest