@SanchosPancho 2024-02-15
I've test your PR and seems something wrong with it. After plan I've got an error:
```

│ Error: kafka: client has run out of available brokers to talk to: 5 errors occurred:
│ 	* EOF
│ 	* EOF
│ 	* EOF
│ 	* EOF
│ 	* EOF
```

Without this patch all works good