@eh-steve 2024-02-19
Yeah I guess if you wanted to test this PR you could locally cherry pick the first commit onto an older master.

I've also released it in the registry under `eh-steve/kafka` version 0.5.5 so you could test with that. The parent master commit for that release is 993494b, which may have already introduced the problem though?