@eh-steve 2024-03-25
> Hey @eh-steve , Thanks for the above optimizations. Does the above reduce the time execution, on regular terraform plan commands for clusters that have tons of ACLs ?

Yeah it should 


> BTW is there any help you need, to push forward with this PR?

Just waiting for an approval and a CI run