@Mongey 2024-04-04
hey @eh-steve thanks for the PR! looks great!
The build snapshots never really works for PR's due to GPG, I generally just ignore that failure. 

I pulled the branch down locally and tested -- it seems to work for me. 
Not sure what's going on with CI - it is quite flakey